+++ src/types/Event.ts
@@ -1,6 +1,7 @@
+import { EventType } from '$api/event';
 import { Item } from 'types/Items';
 
-export type Event = {
+export type TargetEvent = {
   id: number;
   title: string;
   code: string;
@@ -14,3 +15,15 @@ export type Event = {
   missMessage: string;
   missImageUrl: string;
 };
+
+
+export interface EventForm {
+  title: string
+  openAt: string
+  closeAt: string
+  hitMessage: string
+  missMessage: string
+  type: EventType
+  hitImageUrl: string
+  missImageUrl: string
+}
